Make classing wave to arrive instead of just showign up

+++ src/app.rs
@@ -9,7 +9,7 @@
 //
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -192,7 +192,7 @@ impl App {
                     self.score += 1;
 
                     beam.y = -1000;
-                    beam_explosions.push((enemy.x + e_w / 2, enemy.y + e_h / 2));
+                    beam_explosions.push((enemy.x as u32 + e_w / 2, enemy.y as u32 + e_h / 2));
                     play_explosion = true;
 
                     // Type-safe property reset using 'if let'
